wip: implement FastAPI application structure with logging and middleware support
