+++ docs/release-notes.md
@@ -16,5 +16,3 @@ hide:
 
 
 **Full Changelog**: https://github.com/bybatkhuu/module-fastapi-logging/compare/v1.1.2...v2.0.0
-
-
