+++ examples/lifespan.py
@@ -1,0 +1,34 @@
+from collections.abc import AsyncGenerator
+from contextlib import asynccontextmanager
+
+from fastapi import FastAPI
+from logger import logger
+
+from __version__ import __version__
+
+
+@asynccontextmanager
+async def lifespan(app: FastAPI) -> AsyncGenerator[None, None]:
+    """Lifespan context manager for FastAPI application.
+    Startup and shutdown events are logged.
+
+    Args:
+        app (FastAPI, required): FastAPI application instance.
+    """
+
+    logger.info("Preparing to startup...")
+
+    # Add startup code here...
+    logger.success("Finished preparation to startup.")
+    logger.opt(colors=True).info(f"Version: <c>{__version__}</c>")
+
+    yield
+
+    logger.info("Praparing to shutdown...")
+    # Add shutdown code here...
+    logger.success("Finished preparation to shutdown.")
+
+
+__all__ = [
+    "lifespan",
+]
