+++ src/beans_logging_fastapi/config.py
@@ -1,0 +1,9 @@
+from beans_logging.config import LoggerConfigPM
+
+
+class FastAPIConfigPM(LoggerConfigPM):
+    pass
+
+
+class FastAPILoggerConfigPM(LoggerConfigPM):
+    fastapi: dict = {}
